requirements.txt
ansicolors >= 1.0.2
beautifulsoup4 >= 4.3.2
biopython >= 1.65
coloredlogs >= 1.0.1
ftputil >= 2.2
pandas >= 0.16
progress >= 1.2
pyfaidx >= 0.3.4
requests >= 2.7.0
tabulate >= 0.7.5
